Remove sqltestutil
