+++ instrumentation/database/sql/splunksql/test/db_metrics_test.go
@@ -15,16 +15,24 @@
 package test
 
 import (
+	"context"
 	"database/sql"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/sdk/instrumentation"
+	"go.opentelemetry.io/otel/sdk/metric"
+	"go.opentelemetry.io/otel/sdk/metric/metricdata"
+	"go.opentelemetry.io/otel/sdk/metric/metricdata/metricdatatest"
+	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
 
+	splunkotel "github.com/signalfx/splunk-otel-go"
 	"github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql"
-	"github.com/signalfx/splunk-otel-go/instrumentation/internal/sqltestutil"
 )
 
-func TestMetrics(t *testing.T) {
+func TestMetrics(t *testing.T) { //nolint:funlen // the want is big
 	testCases := []struct {
 		driverName       string
 		connectionString string
@@ -43,6 +51,12 @@ func TestMetrics(t *testing.T) {
 	}
 	for _, tc := range testCases {
 		t.Run(tc.driverName, func(t *testing.T) {
+			// arrange
+			ctx := context.Background()
+			reader := metric.NewManualReader()
+			meterProvider := metric.NewMeterProvider(metric.WithReader(reader))
+			defer func() { assert.NoError(t, meterProvider.Shutdown(ctx)) }()
+
 			// instrument: register the fake driver
 			driver := newSimpleMockDriver()
 			connCfg := splunksql.ConnectionConfig{
@@ -54,11 +68,86 @@ func TestMetrics(t *testing.T) {
 				DSNParser: func(string) (splunksql.ConnectionConfig, error) { return connCfg, nil },
 			})
 
-			// execute test
-			sqltestutil.TestMetrics(t, tc.wantPoolName, tc.driverName, "dataSourceName", func(db *sql.DB) {
-				_, err := db.Exec("SELECT 1")
-				require.NoError(t, err)
-			})
+			// create 1 used connection
+			db, err := splunksql.Open(tc.driverName, "dataSourceName", splunksql.WithMeterProvider(meterProvider))
+			require.NoError(t, err)
+			defer func() { assert.NoError(t, db.Close()) }()
+			_, err = db.Exec("SELECT 1")
+			require.NoError(t, err)
+
+			// assert
+			wantPoolAttr := attribute.String("pool.name", tc.wantPoolName)
+			want := metricdata.ScopeMetrics{
+				Scope: instrumentation.Scope{
+					Name:      "github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql",
+					Version:   splunkotel.Version(),
+					SchemaURL: semconv.SchemaURL,
+				},
+				Metrics: []metricdata.Metrics{
+					{
+						Name:        "db.client.connections.usage",
+						Unit:        "{connection}",
+						Description: "The number of connections that are currently in state described by the state attribute",
+						Data: metricdata.Sum[int64]{
+							DataPoints: []metricdata.DataPoint[int64]{
+								{
+									Attributes: attribute.NewSet(
+										wantPoolAttr,
+										attribute.String("state", "used"),
+									),
+									Value: 0,
+								},
+								{
+									Attributes: attribute.NewSet(
+										wantPoolAttr,
+										attribute.String("state", "idle"),
+									),
+									Value: 1,
+								},
+							},
+							Temporality: metricdata.CumulativeTemporality,
+							IsMonotonic: false,
+						},
+					},
+					{
+						Name:        "db.client.connections.max",
+						Unit:        "{connection}",
+						Description: "The maximum number of open connections allowed",
+						Data: metricdata.Sum[int64]{
+							DataPoints: []metricdata.DataPoint[int64]{
+								{
+									Attributes: attribute.NewSet(wantPoolAttr),
+									Value:      0,
+								},
+							},
+							Temporality: metricdata.CumulativeTemporality,
+							IsMonotonic: false,
+						},
+					},
+					{
+						Name:        "db.client.connections.wait_time",
+						Unit:        "ms",
+						Description: "The time it took to obtain an open connection from the pool",
+						Data: metricdata.Sum[int64]{
+							DataPoints: []metricdata.DataPoint[int64]{
+								{
+									Attributes: attribute.NewSet(wantPoolAttr),
+									Value:      0,
+								},
+							},
+							Temporality: metricdata.CumulativeTemporality,
+							IsMonotonic: false,
+						},
+					},
+				},
+			}
+
+			rm := metricdata.ResourceMetrics{}
+			err = reader.Collect(ctx, &rm)
+			require.NoError(t, err)
+			require.Len(t, rm.ScopeMetrics, 1, "should export metrics")
+			got := rm.ScopeMetrics[0]
+			metricdatatest.AssertEqual(t, want, got, metricdatatest.IgnoreTimestamp())
 		})
 	}
 }
