+++ instrumentation/database/sql/splunksql/test/go.mod
@@ -3,11 +3,12 @@ module github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql
 go 1.18
 
 require (
+	github.com/signalfx/splunk-otel-go v1.4.0
 	github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql v1.4.0
-	github.com/signalfx/splunk-otel-go/instrumentation/internal/sqltestutil v1.4.0
 	github.com/stretchr/testify v1.8.2
 	go.opentelemetry.io/otel v1.14.0
 	go.opentelemetry.io/otel/sdk v1.14.0
+	go.opentelemetry.io/otel/sdk/metric v0.37.0
 	go.opentelemetry.io/otel/trace v1.14.0
 	go.uber.org/goleak v1.2.1
 )
@@ -18,10 +19,8 @@ require (
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/kr/text v0.2.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/signalfx/splunk-otel-go v1.4.0 // indirect
 	github.com/signalfx/splunk-otel-go/instrumentation/internal v1.4.0 // indirect
 	go.opentelemetry.io/otel/metric v0.37.0 // indirect
-	go.opentelemetry.io/otel/sdk/metric v0.37.0 // indirect
 	golang.org/x/sys v0.5.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
@@ -30,5 +29,4 @@ replace (
 	github.com/signalfx/splunk-otel-go => ../../../../../
 	github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql => ../
 	github.com/signalfx/splunk-otel-go/instrumentation/internal => ../../../../internal/
-	github.com/signalfx/splunk-otel-go/instrumentation/internal/sqltestutil => ../../../../internal/sqltestutil/
 )
