--- instrumentation/internal/sqltestutil/doc.go
@@ -1,22 +1,0 @@
-// Copyright Splunk Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/*
-Package sqltestutil contains helpers for testing database instrumentations
-build on top of the splunksql instrumentation.
-This package is in a separate module from the instrumentation it tests to
-isolate the dependency of the default SDK and not impose this as a transitive
-dependency for users.
-*/
-package sqltestutil // import "github.com/signalfx/splunk-otel-go/instrumentation/internal/sqltestutil"
